requirements.txt
# Core Dependencies
pandas==2.1.4
numpy==1.26.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Production
gunicorn==21.2.0